tuning. libc is 1-2 seconds faster to load now.
git-svn-id: http://svn.netlabs.org/repos/kbuild/trunk@526 8c291193-6614-0410-a08d-adc6e07170bb

+++ src/gmake/variable.h
@@ -53,8 +53,8 @@ struct variable
     char *name;			/* Variable name.  */
     int length;			/* strlen (name) */
 #ifdef VARIABLE_HASH /* bird */
-    long hash1;                 /* the primary hash */
-    long hash2;                 /* the secondary hash */
+    int hash1;                  /* the primary hash */
+    int hash2;                  /* the secondary hash */
 #endif
     char *value;		/* Variable value.  */
     struct floc fileinfo;       /* Where the variable was defined.  */
